added email validator and fixed live object detection
