+++ app.py
@@ -4,6 +4,7 @@ from poc.songsdetection.views import songs_blueprint
 from poc.spamham.views import spamham_blueprint
 from poc.objectdetection.views import object_detection_blueprint
 from poc.imageclassification.views import image_classification_blueprint
+from poc.emailvalidator.views import email_validator_blueprint
 
 app = Flask(__name__)
 app.config["SECRET_KEY"] = "mysecret"
@@ -22,6 +23,8 @@ app.register_blueprint(songs_blueprint,
                        url_prefix='/songs')
 app.register_blueprint(image_classification_blueprint,
                        url_prefix='/image-classification')
+app.register_blueprint(email_validator_blueprint,
+                       url_prefix='/email-validator')
 
 if __name__ == '__main__':
     app.run(debug=True)
