+++ poc/objectdetection/coco.names
