+++ poc/objectdetection/ssd_mobilenet_v3_large_coco_2020_01_14.pbtxt
