+++ poc/objectdetection/views.py
@@ -1,8 +1,7 @@
 from flask import render_template, Response, jsonify, render_template, Blueprint
 from flask_restful import Api
-from poc.songsdetection.camera import *
-df1 = music_rec()
-df1 = df1.head(15)
+from poc.objectdetection.camera import *
+
 
 object_detection_blueprint = Blueprint(
     'objectdetection', __name__, template_folder='templates/')
@@ -19,11 +18,6 @@ def object_detection():
 def video_feed_for_object():
     return Response(generate_webcam(VideoCameraForObject()),
                     mimetype='multipart/x-mixed-replace; boundary=frame')
-
-
-@object_detection_blueprint.route('/t')
-def gen_table():
-    return df1.to_json(orient='records')
 
 
 def generate_webcam(camera):
